placessidebar: Support smart bookmarking of disk volumes/mounts
Treat bookmarks for volumes/mounts specially. Currently, when the user
bookmarks the root directory/mount point of a volume, the bookmark is
stored as a regular file:/// uri of the current mount point in the
filesystem. In other words, the bookmarked location is treated as a
regular directory. This leads to the obvious problem that when the
user selects the same bookmark when the volume is no longer mounted,
the bookmark fails to open with a generic error message.

We make bookmarking of volume/mounts smarter by storing the bookmark as
the volume's unique id in the form "volume:///<Volume UUID>" - using an
artificial url scheme represented by "volume:///" - instead of the
mount path in the bookmarks file. When populating the bookmarks in
the sidebar, on recognizing the url scheme as one representing a volume
we handle the mounting and unmounting of the volume the same way as we
handle volumes/drives that are listed by default.