GDK W32: keep window size a multiple of scaling factor
This prevents window from bobbing slightly when resized.
The logic for rounding the size up and down is by no means
set in stone, and might be changed later.

We need an extra RECT in the window impl to keep track of
the window size changing over time (can't use GdkWindow width/height
or impl unscaled_with/height for this - they suffer too many
transformations).