Checkbutton: Fix moving the focus with the arrow keys for check boxes
Currently when the focus is on a check box (so a check button without a
group), using the arrow keys to move the focus instead "ticks"/checks the
check box instead of moving the focus in the desired direction. The focus
is kept locked in the check button until Tab is used.

To fix this, this commit checks that the check button is a check box, in
which case focus movements will be done normally without triggering the
radio button behavior that "moves" the pressed radio button within this
group of radio buttons.
There was also another issue that prevented from correctly using the
arrow keys to move the focus, which happened when there was a widget
with a radio button for example on its right, where the radio button
would get skipped when moving the focus from the widget to the radio
button if it isn't activated, and the next widget in this direction
would get focused instead. The offending code was just removed, as
reading it I just can't make any sense of its original behavior: to me
that code meant "if we're moving the focus to a check button which isn't
the activated radio button in the group, then skip the focus chain
(i.e. return false)".