Modify the entry value by checking the preedit cursor position
In the case of Korean that has not been committed by inputting one character, there is a problem that the selected characters are not deleted when a key is input with a string selected. In this case, the entered key is displayed behind the selected character, making it awkward. Therefore, the position of the preedit cursor is checked in gtk_entry_preedit_changed_cb, which receives key input, and if it is 0, the selected strings are replaced with null strings.