Correctly detect usability of glBlitFramebuffer on GLES
The glBlitFramebuffer function is used to copy an image from one GL
framebuffer to another.  This function is defined by several GL
versions and extensions:

- by the GL extension GL_EXT_framebuffer_blit, as glBlitFramebufferEXT
- by the GL extension GL_ARB_framebuffer_object, as glBlitFramebuffer
- by GL 3.0, as glBlitFramebuffer
- by the GLES extension GL_NV_framebuffer_blit, as glBlitFramebufferNV
- by GLES 3.0, as glBlitFramebuffer

All of these are effectively equivalent, and therefore libepoxy treats
glBlitFramebuffer, glBlitFramebufferEXT, and glBlitFramebufferNV as
aliases for each other.

The GL_ANGLE_framebuffer_blit extension is different; it defines a
function glBlitFramebufferANGLE which has a limited subset of the
functionality of glBlitFramebuffer.  In fact, that subset would be
adequate for GDK's current usage; but libepoxy doesn't treat the two
functions as aliases, so in order to use the ANGLE extension, GDK
would need to call glFramebufferBlitANGLE instead.

(The message for commit af66faf6048977c6c27e17c95f6c3c47440904d9,
which introduced the check for GL_ANGLE_framebuffer_blit, suggests
that the code required GLES 3.0 for other reasons, so the fact that
this check seemed to work was a coincidence.)

Therefore:
- if running on GLES 3.0, enable use of glFramebufferBlit.
- if running on GLES 2.0 with the GL_NV_framebuffer_blit extension,
  enable use of glFramebufferBlit.
- do *not* enable use of glFramebufferBlit solely based on the
  GL_ANGLE_framebuffer_blit extension.

This should provide better performance on common platforms, and
possibly avoid crashes on some uncommon platforms.