gdk_cairo_draw_from_gl: detect opaque textures on GLES
If the input to gdk_cairo_draw_from_gl is opaque (doesn't contain an
alpha channel), then we don't need to worry about blending it with the
background.

In the case of a texture, GLES doesn't support
glGetTexLevelParameteriv, so there is no way to directly check whether
a texture has an alpha channel.  However, we can check by binding the
texture to a temporary framebuffer and then querying the
GL_ALPHA_BITS.  (It is assumed in any case that it is *possible* to
bind the supplied texture to a framebuffer; the software fallback case
later in this function relies on doing so.)

The nuisance of performing this query should be outweighed by the
benefit of avoiding unnecessary and expensive blending.