GtkGLArea: prefer textures over renderbuffers if we can't blit
GtkGLArea can use either a texture or a renderbuffer for its temporary
working framebuffer, which must later be transferred to the actual
back buffer of the window.

Copying a renderbuffer to the back buffer (using glBlitFramebuffer)
can be fast and efficient, if the driver/hardware supports that
operation.  Copying a texture to the back buffer (by applying it as a
texture to a full-screen rectangle) is slightly less efficient.
Therefore, GTK would previously use a texture only if an alpha channel
was requested.

However, if glBlitFramebuffer is unavailable (GL 2.0 without the
GL_EXT_framebuffer_blit extension, or GLES 2.0 without the
GL_NV_framebuffer_blit extension), then using a texture makes more
sense.

This logic should probably be kept in sync with the logic in GDK
(gdk_gl_context_check_extensions), but nothing should break if the two
are out of sync.

If we're drawing to a texture and we don't want an alpha channel, the
texture should be created with an internal format of GL_RGB8 rather
than GL_RGBA8; this will tell gdk_cairo_draw_from_gl that the image
doesn't need to be blended with the background.

(I'm not sure that the use of GL_BGRA instead of GL_RGBA, in the
following lines, has any effect at all, but it doesn't look like it
harms anything.)