gdk_cairo_draw_from_gl: fix red/blue swapping on GLES
In the "software fallback" case, when we are trying to draw a
renderbuffer or texture onto a cairo surface and we don't have an
accelerated way to do so with GL, we must read the pixel data from GL
into a cairo image.

In the case of desktop GL, we can read the image in cairo's native
format (CAIRO_FORMAT_ARGB32 or CAIRO_FORMAT_RGB24) since both of these
are equivalent to "GL_BGRA, GL_UNSIGNED_INT_8_8_8_8_REV".

In the case of GLES, the only format that glReadPixels is required to
support is "GL_RGBA, GL_UNSIGNED_BYTE".  Some GLES implementations
(including Mesa and ANGLE) allow using "GL_BGRA, GL_UNSIGNED_BYTE"
instead.  But there doesn't appear to be a way for the implementation
to advertise this capability.  (The GL_EXT_read_format_bgra extension
appears to *permit* the capability, but not to *require* it unless
BGRA is the actual internal storage format.)

Currently we're assuming that if we're using GLES on Windows, we must
be using ANGLE and therefore BGRA format must be supported.  (This
assumption seems dubious.)

But in any case, if we don't know that the implementation supports the
non-standard BGRA format, we must use RGBA; we then need to swap the
channels to the order that cairo expects.