gesturestylus: honor `GtkGestureSingle:button` property
Use the `GtkGestureSingle:button` property to determine which
buttons count for stylus contact. Press/release events for
other buttons will not be consumed by the gesture.

With the property's default value of 0, any button press
will count as contact, which is approximately the historical
behavior, to ensure some measure of backwards compatibility.

Most users will want to set the property to 1, so that only
button 1 registers as contact.