builder: Accept standard C type names
We are much more likely to use standard C types
in our code nowadays, and it is a bit of a
dissonance to have to use the g-names when
referring to the same types in ui files.

This adds code to our type name resolver
to map the standard C numeric type names
to their glib equivalents.

Tests included.