Make enabling kinetic_scrolling not break scrolling
The GesturePan should only be in a capture phase for specific conditions.
These conditions are checked in check_attach_pan_gesture and here the GesturePan is also enabled if need be.
Overriding this by always setting the phase to capture when kinetic-scrolling is enabled leads to broken scrolling for whichever orientation the GestureDrag may have set at that point.

Fixes #5759