wayland: Update to xdg-decoration protocol
Currently gtk uses the out of date and non-standardized
server-decoration protocol from kde. This commit replaces the
implementation of this old server-decoration protocol with an
implementation of the xdg-decoration protocol standardized in the
wayland-protocols repository.

Notable changes include:

- New private GDK API to communicate decoration state changes to GTK

- Deprecation of gdk_wayland_display_prefers_ssd(). The new protocol
doesn't provide global state, so this function is deprecated and always
returns false.

- Renaming of gdk_wayland_window_announce_ssd() to
gdk_wayland_window_request_ssd(). The server is not required to
honor requests for ssd. It is however required to honor CSD if the
xdg-decoration protocol object is destroyed. so
gdk_wayland_window_announce_csd() is still an accurate name.

Signed-off-by: Dallas Strouse <dastrouses@gmail.com>
Co-authored-by: Isaac Freund <mail@isaacfreund.com>