gtklabel: Queue a draw after changing the focused link
When tabbing between three widgets, a button, a label with hyperlink, and
another button, the hyperlink doesn’t get its focus rectangle rendered
when keyboard focus is on it — unless the user then mouses over it. The
label is non-selectable.

It looks like this might have been because the label was not queued for
redraw after changing the focused link
(`label->select_info->selection_anchor`), so I tried this patch.

This patch doesn’t actually fix the issue though. So obviously I’m
missing something.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>