boxlayout: Measure a single variable-size child directly
Since commit 76c46739447023368db4c3e6d008b52ff4b249e0
"boxlayout: Fix broken min-size-for-opposite-size", GtkBoxLayout runs a
binary search to determine its minimum size along the opposite
orientation, trying different sizes and seeing if its children fit the
provided for_size constraint along the natural orientation. This is only
done if any of the box's children are not constant-size.

A notable source of non-constant-size-ness in a widget tree is a
GtkLabel with wrap = true, which runs a single Pango layout to measure
its height-for-width, and its own binary search, each time doing a Pango
layout, in case it has to measure its width-for-height. So if a
wrappable label is located in a horizontal box, there will be two levels
of binary searches performed. Moreover, the amount of nested binary
searches gets amplified if the label is wrapped into several layers of
boxes of alternating orientations, which causes a huge number of Pango
layouts to be performed.

This has been found to be a source of *major* performance issues in
Fractal [0] and Paper Plane [1] chat apps.

[0]: https://gitlab.gnome.org/GNOME/fractal/-/issues/1183
[1]: https://github.com/paper-plane-developers/paper-plane/issues/379

Partly address the issue by noticing that the common case is actually
the box having a single non-constant-size child, and a number of
constant-size children; for example, this can be a wrappable label with
the message text (height-for-width), sender's avatar (constant-size),
and a non-wrappable label with the date/time the message was sent (also
constant-size). This is also true of the many layers of outer boxes;
where the inner box is the only non-constant-size child.

In this simple and common case, we can ask the single non-constant-size
child for its minimum size in the opposite orientation directly, without
running a binary search. This cuts down on the number of Pango layouts
dramatically.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>