flowbox: Report constant-size request mode when empty
Constant size widgets are much easier for the layout machinery to deal
with; in particular this can help GtkBoxLayout to hit a faster measuring
path.

This is prompted by the Fractal app using a GtkFlowBox to display emoji
reactions to a chat message. For most messages, there are no reactions,
so the GtkFlowBox could report being constant-size, which would enable
the containing GtkBox'es to do their layout much faster.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>