a11y: Support indicating the minimum increment for range like controls
This introduces the GTK_ACCESSIBLE_PROPEETY_MINIMUM_INCREMENT, which allows
indicating the expected minimum increment of a range like control.

This allows screen readers to present the control's value in a much nicer
format, for example, respecting the number of digits after the decimal point.

This is exactly the behavior which Orca is doing, and fixes its reading of
scales with fractional values.