label: Allow focusing without being selectable
Marking a non-selectable label as focusable is useful when a label
that isn't associated with a control should be "visible" to screen
readers.

Currently code that needs this has to resort to hacks[0], make
those unnecessary by respecting the :focusable property.

As widgets are not focusable by default, this shouldn't affect
existing users.

[0]
https://gitlab.gnome.org/GNOME/gnome-control-center/-/blob/45.0/panels/common/cc-list-row-info-button.ui#L26