a11y: Publish API to update platform state
API for updating the platform state was private so far,
i.e. not available for custom GtkAccessible implementations.

This e.g. didn't allow custom implementations to update the
focused state, which is essential for assistive technology
to be able to track focus and announce the currently focused
item.

Therefore, move the GtkAccessiblePlatformChange enum and
the declaration of gtk_accessible_platform_changed to the
public header.

Adjust gtk_at_spi_context_platform_change to not only work
for GtkWidget, but any GtkAccessible.

Fixes: #6272