gtk: Added "popup-shown" property to Gtk.DropDown
This read-only property can be used to determine if the popup is
currently open or closed. Using "notify::popup-shown" it's also possible
to be notified when the drop down opens/closes using a callback.

Gtk.ComboBox has this property and I find it useful to have it on
Gtk.DropDown too.

Signed-off-by: Florian "sp1rit" <sp1rit@disroot.org>