gtk3: fix for gtk3-demo-application
The menus of the demo application were misnamed and
the app menu function not used correctly

As mentioned in #6524 the gtk3 demo application was also
showing the wrong menus and since the file handling menu
was moved to the appmenu it was missing from system that
have "gtk-shell-shows-app-menu" disabled.

The about action is changed from a window to an application
action as is usually the default behavior.