wayland: Support relative dials on tablet pads
This is mostly copy/paste from the ring/strip support. It adds a new
event type GDK_PAD_DIAL but otherwise works along the same lines as
rings and strips.

Notably, the event value is normalized to be a multiple of logical wheel
clicks, i.e. the value 1.0 means one wheel click, 0.3 means a third of a
click etc.

In gtk4-demo Paint the dial is mapped to change the brush size.