quartz: add automagic uti<->mime conversion for clipboard
The current version does not convert the mime type "text/html"
to the MacOS universal type identifier "public.html" on the clipboard.
This patch checks if the mime type is known in the Apple domain and then
uses the proposed UTI type. The conversion back works the same
way. So with this patch not only "text/html" is converted to the
correct uti but also other mime types. For example "image/svg+xml"
is translated to public.svg-image. If the mime type is not known,
then the behaviour is as before, i.e. the type is simply announced
on the clipboard and a dynamic uti is created by MacOS.

With this patch applications like Apple Email, Notes or Keynote can
exchange data with gtk applications. I checked this with pspp where
html, svg and image data is available on the clipboard. The problem
is discussed and described in the pspp bug tracking:

https://savannah.gnu.org/bugs/?65288