wayland: Fix reported monitor positions for xdg_output v3+
In version 3 and above of the xdg_output protocol, the xdg_output.done
event is no longer sent.  Instead, wl_output.done is sent again after
the xdg_output events have all been sent.

Unfortunately this means that it's a bit more complicated to discover in
the wl_output.done handler if we are handling the completion of the
wl_output properties or the xdg_output properties.